+++ app/test/e2e/queue-execution-create.spec.js
@@ -1,4 +1,4 @@
-/* eslint-disable no-unused-vars,no-undef,no-await-in-loop */
+/* eslint-disable no-await-in-loop */
 const nock = require('nock');
 const chai = require('chai');
 const amqp = require('amqplib');
@@ -13,9 +13,8 @@ const sleep = require('sleep');
 const { getTestServer } = require('./test-server');
 
 chai.use(chaiMatch);
-const should = chai.should();
+chai.should();
 
-let requester;
 let rabbitmqConnection = null;
 let channel;
 
@@ -64,7 +63,7 @@ describe('EXECUTION_CREATE handling process', () => {
         const dataQueueStatus = await channel.checkQueue(config.get('queues.data'));
         dataQueueStatus.messageCount.should.equal(0);
 
-        requester = await getTestServer();
+        await getTestServer();
     });
 
     beforeEach(async () => {